requirements.txt
autopep8==1.5.4
Django==3.1
djangofilter==2.3.0
djangorestframework==3.11.1
markdown==3.2.2
pylint==2.6.0